move func to generic slices testutil file
`VerifySameElements` does not cater to or have any specific dependency
on resources, so it doesn't make sense to have in
`testutils/resource_utils.go`
