+++ testutils/resource_utils.go
@@ -63,22 +63,6 @@ func VerifySameResourceNames(tb testing.TB, actual, expected []resource.Name) {
 	test.That(tb, newSortedResourceNames(actual), test.ShouldResemble, newSortedResourceNames(expected))
 }
 
-// VerifySameElements asserts that two slices contain the same elements without
-// considering order.
-func VerifySameElements[E cmp.Ordered](tb testing.TB, actual, expected []E) {
-	tb.Helper()
-
-	actualSorted := make([]E, len(actual))
-	copy(actualSorted, actual)
-	expectedSorted := make([]E, len(expected))
-	copy(expectedSorted, expected)
-
-	slices.Sort(actualSorted)
-	slices.Sort(expectedSorted)
-
-	test.That(tb, actualSorted, test.ShouldResemble, expectedSorted)
-}
-
 // ExtractNames takes a slice of resource.Name objects
 // and returns a slice of name strings for the purposes of comparison
 // in automated tests.
